move save track related components to separate folder
