+++ src/frontend/screens/SaveTrack/SaveTrackScreen.tsx
@@ -7,22 +7,22 @@ import {
   StyleSheet,
   View,
 } from 'react-native';
-import {DiscardModal} from '../../../sharedComponents/DiscardModal';
-import {BottomSheet} from '../../../sharedComponents/BottomSheet/BottomSheet';
-import PhotoIcon from '../../../images/camera.svg';
-import DetailsIcon from '../../../images/details.svg';
-import TrackIcon from '../../../images/Track.svg';
+import {DiscardModal} from '../../sharedComponents/DiscardModal';
+import {BottomSheet} from '../../sharedComponents/BottomSheet/BottomSheet';
+import PhotoIcon from '../../images/camera.svg';
+import DetailsIcon from '../../images/details.svg';
+import TrackIcon from '../../images/Track.svg';
 import {defineMessages, useIntl} from 'react-intl';
-import {Text} from '../../../sharedComponents/Text';
-import {TrackDescriptionField} from './saveTrack/TrackDescriptionField';
-import {useBottomSheetModal} from '../../../sharedComponents/BottomSheetModal';
-import DiscardIcon from '../../../images/delete.svg';
-import ErrorIcon from '../../../images/Error.svg';
-import {usePersistedTrack} from '../../../hooks/persistedState/usePersistedTrack';
-import {useNavigationFromRoot} from '../../../hooks/useNavigationWithTypes';
+import {Text} from '../../sharedComponents/Text';
+import {TrackDescriptionField} from './TrackDescriptionField';
+import {useBottomSheetModal} from '../../sharedComponents/BottomSheetModal';
+import DiscardIcon from '../../images/delete.svg';
+import ErrorIcon from '../../images/Error.svg';
+import {usePersistedTrack} from '../../hooks/persistedState/usePersistedTrack';
+import {useNavigationFromRoot} from '../../hooks/useNavigationWithTypes';
 import {CommonActions, useFocusEffect} from '@react-navigation/native';
-import {SaveTrackButton} from './saveTrack/SaveTrackButton';
-import Close from '../../../images/close.svg';
+import {SaveTrackButton} from './SaveTrackButton';
+import Close from '../../images/close.svg';
 
 export const SaveTrackScreen = () => {
   const navigation = useNavigationFromRoot();
